Updated test namespaces, 'phpunit.xml' suite paths, and 'phpstan.neon' ignore path for the restructured test tree.
