+++ tests/Drupal/Tests/Driver/Kernel/Core/CoreSystemMethodsKernelTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\Tests\Driver\Kernel;
+namespace Drupal\Tests\Driver\Kernel\Core;
 
 use Drupal\Driver\Core\Core;
 use Drupal\KernelTests\KernelTestBase;
