+++ tests/Drupal/Tests/Driver/Kernel/Core/Field/DatetimeHandlerKernelTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\Tests\Driver\Kernel;
+namespace Drupal\Tests\Driver\Kernel\Core\Field;
 
 use Drupal\datetime\Plugin\Field\FieldType\DateTimeItem;
