+++ tests/Drupal/Tests/Driver/Unit/Core/CoreFieldMethodsTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\Tests\Driver\Core;
+namespace Drupal\Tests\Driver\Unit\Core;
 
 use Drupal\Core\Entity\EntityFieldManagerInterface;
 use Drupal\Core\Field\BaseFieldDefinition;
