+++ tests/Drupal/Tests/Driver/Unit/Core/CoreTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\Tests\Driver\Core;
+namespace Drupal\Tests\Driver\Unit\Core;
 
 use Drupal\Core\CronInterface;
 use Drupal\Driver\Core\Core;
