+++ tests/Drupal/Tests/Driver/Unit/Core/Field/FileHandlerTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\Tests\Driver\Core\Field;
+namespace Drupal\Tests\Driver\Unit\Core\Field;
 
 use Drupal\Core\DependencyInjection\ContainerBuilder;
 use Drupal\Driver\Core\Field\FileHandler;
