+++ tests/Drupal/Tests/Driver/Unit/Core/FieldHandlerLookupTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Drupal\Tests\Driver\Core;
+namespace Drupal\Tests\Driver\Unit\Core;
 
 use Drupal\Core\DependencyInjection\ContainerBuilder;
 use Drupal\Core\Entity\EntityFieldManagerInterface;
